# web/app/lib/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

/**
 * Article
 */
export type Article = {
    /**
     * Kind
     */
    kind?: 'article';
    /**
     * Title
     */
    title: string;
    /**
     * Alt Title
     */
    alt_title?: Array<string>;
    /**
     * Summary
     */
    summary?: string;
    /**
     * Sections
     */
    sections?: Array<Section>;
    /**
     * Links
     */
    links?: {
        [key: string]: Array<string>;
    };
};

/**
 * GenerateArticleRequest
 */
export type GenerateArticleRequest = {
    /**
     * Lang
     */
    lang?: string;
    /**
     * Context
     */
    context?: string | Array<string> | 'auto' | null;
    /**
     * Style Inst
     */
    style_inst?: string | null;
    /**
     * Streaming
     */
    streaming?: number | boolean;
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * ProtoArticle
 */
export type ProtoArticle = {
    /**
     * Kind
     */
    kind?: 'proto_article';
    /**
     * Title
     */
    title: string;
    /**
     * Alt Title
     */
    alt_title?: Array<string>;
};

/**
 * ProtoWithContextResponse
 */
export type ProtoWithContextResponse = {
    /**
     * Proto
     */
    proto: ({
        kind: 'proto_article';
    } & ProtoArticle) | ({
        kind: 'article';
    } & Article);
    /**
     * Context
     */
    context: Array<string>;
};

/**
 * ProtoWithRelatedResponse
 */
export type ProtoWithRelatedResponse = {
    /**
     * Proto
     */
    proto: ({
        kind: 'proto_article';
    } & ProtoArticle) | ({
        kind: 'article';
    } & Article);
    /**
     * Related Articles
     */
    related_articles: Array<Article>;
};

/**
 * Section
 */
export type Section = {
    /**
     * Title
     */
    title: string;
    /**
     * Content
     */
    content: string;
};

/**
 * ValidationError
 */
export type ValidationError = {
    /**
     * Location
     */
    loc: Array<string | number>;
    /**
     * Message
     */
    msg: string;
    /**
     * Error Type
     */
    type: string;
};

export type RootData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/';
};

export type RootResponses = {
    /**
     * Successful Response
     */
    200: unknown;
};

export type GenerateArticleData = {
    body: GenerateArticleRequest;
    path: {
        /**
         * Title
         */
        title: string;
        /**
         * Space Name
         */
        space_name: string;
    };
    query?: never;
    url: '/spaces/{space_name}/articles/{title}/generate';
};

export type GenerateArticleErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type GenerateArticleError = GenerateArticleErrors[keyof GenerateArticleErrors];

export type GenerateArticleResponses = {
    /**
     * Successful Response
     */
    200: Article;
};

export type GenerateArticleResponse = GenerateArticleResponses[keyof GenerateArticleResponses];

export type GetArticleData = {
    body?: never;
    path: {
        /**
         * Title
         */
        title: string;
        /**
         * Space Name
         */
        space_name: string;
    };
    query?: never;
    url: '/spaces/{space_name}/articles/{title}';
};

export type GetArticleErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type GetArticleError = GetArticleErrors[keyof GetArticleErrors];

export type GetArticleResponses = {
    /**
     * Successful Response
     */
    200: Article;
};

export type GetArticleResponse = GetArticleResponses[keyof GetArticleResponses];

export type GetArticleProtoData = {
    body?: never;
    path: {
        /**
         * Title
         */
        title: string;
        /**
         * Space Name
         */
        space_name: string;
    };
    query?: never;
    url: '/spaces/{space_name}/articles/{title}/proto';
};

export type GetArticleProtoErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type GetArticleProtoError = GetArticleProtoErrors[keyof GetArticleProtoErrors];

export type GetArticleProtoResponses = {
    /**
     * Response Get Article Proto
     * Successful Response
     */
    200: ProtoArticle | Article;
};

export type GetArticleProtoResponse = GetArticleProtoResponses[keyof GetArticleProtoResponses];

export type GetArticleProtoWithContextData = {
    body?: never;
    path: {
        /**
         * Title
         */
        title: string;
        /**
         * Space Name
         */
        space_name: string;
    };
    query?: never;
    url: '/spaces/{space_name}/articles/{title}/proto/context';
};

export type GetArticleProtoWithContextErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type GetArticleProtoWithContextError = GetArticleProtoWithContextErrors[keyof GetArticleProtoWithContextErrors];

export type GetArticleProtoWithContextResponses = {
    /**
     * Successful Response
     */
    200: ProtoWithContextResponse;
};

export type GetArticleProtoWithContextResponse = GetArticleProtoWithContextResponses[keyof GetArticleProtoWithContextResponses];

export type GetArticleProtoWithRelatedData = {
    body?: never;
    path: {
        /**
         * Title
         */
        title: string;
        /**
         * Space Name
         */
        space_name: string;
    };
    query?: never;
    url: '/spaces/{space_name}/articles/{title}/proto/related';
};

export type GetArticleProtoWithRelatedErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type GetArticleProtoWithRelatedError = GetArticleProtoWithRelatedErrors[keyof GetArticleProtoWithRelatedErrors];

export type GetArticleProtoWithRelatedResponses = {
    /**
     * Successful Response
     */
    200: ProtoWithRelatedResponse;
};

export type GetArticleProtoWithRelatedResponse = GetArticleProtoWithRelatedResponses[keyof GetArticleProtoWithRelatedResponses];

export type ClientOptions = {
    baseUrl: 'http://localhost:8000' | (string & {});
};